fix tv shows released out of order
Our protection against re-downloading things we've already downloaded is
to check the current `S##E##` against the last episode we downloaded.
This is to prevent shows we start following in season 8 from downloading
seasons 1-7 (which we fill in other ways).

This creates a problem if the TV show is accidentally released
out-of-order, say as part of a full-season release. We won't get any of
the episodes in the gap between the mis-ordered episodes, even though
they're right there as part of the releases.

But, we also check to make sure we haven't downloaded the individual
episode more than once. So, what we really want to do is to make sure
that we don't download an episode older than the very first episode we
started tracking with this script. That way we don't download ancient TV
show episodes we've already seen, and we do download TV shows who are
released out of order.