detect when torrent is not complete
When Transmission has a magnet, it has to fetch a torrent file. Until
that time, it keeps a pseudo-torrent file in its torrents directory. If
I accidentally move the pseudo-torrent file into the Curator database,
very weird errors happen. So now I check for the right info section
before trying to process the torrent.